Move common behavior to abstract strategy class
